- ContentsManager 정의
